+++ seen-common/seen-api/src/main/java/com/songchi/seen/apply/service/UserApplyService.java
@@ -12,21 +12,21 @@ import java.util.Set;
  * @since 2023-01-01
  */
 public interface UserApplyService {
-    List<Integer> appliedUserIdToApplyIdByPage(Integer appliedUserId, int current, int size);
+    List<Integer> appliedUserIdToApplyIdByPage(Long appliedUserId, int current, int size);
 
-    List<Integer> applyUserIdToApplyIdByPage(Integer applyUserId, int current, int size);
+    List<Integer> applyUserIdToApplyIdByPage(Long applyUserId, int current, int size);
 
-    Map<Integer, Integer> appliedUserIdToId(Set<Integer> appliedUserId, Integer userId);
+    Map<Long, Integer> appliedUserIdToId(Set<Long> appliedUserId, Long userId);
 
     Map<Integer, Integer> idToTextId(Set<Integer> ids);
 
-    Map<Integer, Integer> idToApplyUserId(Set<Integer> ids);
+    Map<Integer, Long> idToApplyUserId(Set<Integer> ids);
 
-    Map<Integer, Integer> idToAppliedUserId(Set<Integer> ids);
+    Map<Integer, Long> idToAppliedUserId(Set<Integer> ids);
 
     Map<Integer, LocalDateTime> idToCreateTime(Set<Integer> ids);
 
     Map<Integer, LocalDateTime> idToApplyTime(Set<Integer> ids);
 
-    Integer set(Integer userId, Integer textId, Integer appliedUserId);
+    Integer set(Long userId, Integer textId, Long appliedUserId);
 }
