+++ seen-micro-database/seen-school/src/main/java/com/songchi/seen/school/po/SchoolPO.java
@@ -39,5 +39,5 @@ public class SchoolPO {
     /**
      * 更新用户
      */
-    private Integer updateUser;
+    private Long updateUser;
 }
